Move around the DictNode to new directory
