+++ opencog/nlp/lg-dict/LGDictNode.cc
