+++ opencog/nlp/lg-dict/LGDictNode.h
